Default to 'expandtab' for small indents
For files with small indents and little nesting, confirming the value of
'expandtab' can be hard, which can result in consulting the maximum
number of neighbors without success.  Yaml is particularly bad about
this problem, which is why we currently special case it.  This is an
attempt at a more general solution.