Use short option names in non-verbose :Sleuth output
In the worst case scenario, when every option is set to a value with the
maximum practical length, The :setlocal message displayed by :Sleuth can
reach 128 characters.  Using the short option names reduces this to 73
characters.  So with a tiny sacrifice to readability, we can eliminate
the "Press ENTER" prompt in all but the narrowest of terminals.