Detect 'filetype' from EditorConfig
If we're doing this for modelines, we may as well extend it to
EditorConfig.  I do occasionally encounter bespoke file naming schemes
that I don't want to add to my vimrc.

The "." normalization is so that "." can be used to set an empty
filetype, and "unset." can be used to set a filetype of "unset", in the
unlikely event someone has reason to use one of those values.