Allow overriding problematic .editorconfig
Usage:

    let g:sleuth_editorconfig_overrides = {
          \ expand('~/bad/repo/.editorconfig'): '.editorconfig.mine'}

The value can be a relative or absolute path, or an empty string to
disable the file without replacing it.  On Windows, forward slashes must
be used.

This is a solution to the problem of a repository including a poorly
considered .editorconfig that can't easily be changed.  Leaving
undocumented because I don't know if this is actually a problem in
practice.

References: https://github.com/tpope/vim-sleuth/issues/79