Set 'filetype' from interactive :Sleuth invocations
I thought the displayed output for this would be complicated to
implement since it needs to happen well after the option is set, but
turns out our idempotent handling of unsafe options works for 'filetype'
too.  I was also worried about recursive invocation, but guarding
against that in a <sfile> check seems adequate.