Try harder to detect local 'nomodeline'
The goal here is to skip modelines if 'nomodeline' is set for the
current buffer (e.g., the ftplugin disables it), but not if it was set
globally (e.g., the system vimrc sets it).  The existing logic
successfully rejects `:setglobal nomodeline`, but fails to reject
`:set nomodeline`, since that sets both the local and global option.  So
lets also reject any local option that was set in the exact same place
as the global option.

All of this can be avoided if you leave 'modeline' enabled and instead
`:set modelines=0`.  Thus this is my recommended way to disable
modelines.